Prepare for REDIS
I need to figure out how to handle requests to specific URLs only. I
want to receive the POSTs, but leave all static file handling to
connect. And I need to load stuff from REDIS and populate the fields in
index.html when a request goes to "/123" or the like with "123" being
the REDIS key for a previously saved paste.