Use urlrouter to specify routes
Handling special URLs is easier with urlrouter. I need a way to grab url
path parameters like "/:id" to load previously saved pastes, and I'm too
lazy to do it all by myself.

The only real problem now is putting the loaded paste to the frontend
when a user opens a paste, like "/42".